Chore: Have the separate view file for manage column popup
