+++ lib/cm_admin/view_helpers.rb
@@ -8,6 +8,7 @@ module CmAdmin
     include ColumnFieldHelper
     include FieldDisplayHelper
     include FilterHelper
+    include ManageColumnPopupHelper
     include ActionView::Helpers::FormTagHelper
     include ActionView::Helpers::TagHelper
 
@@ -69,77 +70,5 @@ module CmAdmin
         concat " " + column_path.to_s.gsub('/', '_').humanize
       end
     end
-
-    def manage_column_pop_up(klass)
-      tag.div class: "modal fade form-modal table-column-modal", id: "columnActionModal", role: "dialog", aria: {labelledby: "exampleModalLabel"} do
-        tag.div class: "modal-dialog", role: "document" do
-          tag.div class: "modal-content" do
-            tag.div do
-              concat manage_column_header
-              concat manage_column_body(klass)
-              concat manage_column_footer
-            end
-          end
-        end
-      end
-    end
-
-    def manage_column_header
-      tag.div class: "modal-header" do
-        tag.button type: "button", class: "close", data: {dismiss: "modal"}, aria: {label: "Close"} do
-          tag.span "X", aria: {hidden: "true"}
-        end
-        tag.h5 "Manage columns", class: "modal-title", id: "exampleModalLabel"
-      end
-    end
-
-    def manage_column_body(klass)
-      tag.div class: "modal-body" do
-        tag.div class: "columns-list" do
-          klass.available_fields[:index].each do |column_name|
-            concat manage_column_body_list(column_name)
-          end
-        end
-      end
-    end
-
-    def manage_column_body_list(column_name)
-      tag.div class: "column-item" do
-        concat manage_column_dragger(column_name.field_name)
-        concat manage_column_item_name(column_name)
-        concat manage_column_item_pointer(column_name.field_name)
-      end
-    end
-
-    def manage_column_dragger(default_column_name)
-      return if default_column_name == :id
-      tag.div class: "dragger" do
-        tag.i class: "fa fa-bars bolder"
-      end
-    end
-
-    def manage_column_item_name(column_name)
-      tag.div class: "column-item__name" do
-        tag.p column_name.field_name.to_s.gsub('/', '_').humanize
-      end
-    end
-
-    def manage_column_item_pointer(default_column_name)
-      tag.div class: "column-item__action #{"pointer" if default_column_name != :id}" do
-        tag.i class: "fa #{default_column_name == :id ? "fa-lock" : "fa-times-circle"} bolder"
-      end
-    end
-
-    def manage_column_footer
-      tag.div class: "modal-footer" do
-        tag.button type: "button", class: "gray-border-btn", data: {dismiss: "modal"} do
-          "Close"
-        end
-        tag.button type: "button", class: "cta-btn" do
-          "Save"
-        end
-      end
-    end
-
   end
 end
